% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplots.R
\name{corplots}
\alias{corplots}
\title{Computes all three measures of correlation and adds scatter plots}
\usage{
corplots(dt, group)
}
\description{
Computes all three measures of correlation and adds scatter plots
}
\keyword{internal}
